fix: Decimal serialization error (#102)
The change from Float to Numeric in CitiBuy model tables caused a 
serialization error to be thrown by 
ContractManagement.update_contract_list(). This commit fixes that issue 
by reverting them back to Float.

closes: #101

+++ app/src/dgs_fiscal/systems/citibuy/models/base.py
@@ -4,7 +4,7 @@ from sqlalchemy import (
     Column,
     Integer,
     String,
-    Numeric,
+    Float,
     DateTime,
     ForeignKey,
     ForeignKeyConstraint,
